requirements.txt
playwright
beautifulsoup4
pandas
openpyxl
streamlit-authenticator>=0.3.2
pyyaml
bcrypt
